moved unused nodes to archive folder and reorganized nodes into relevant directories
