+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/action/remove_passed_goals_and_cubes_action.hpp
