+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/action/snap_in_collision_goals_action.hpp
