+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/archive/compute_search_goals_condition.hpp
