+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/archive/get_first_goal_condition.hpp
