+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/archive/goals_empty_condition.hpp
