+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/condition/ar_tag_detected_condition.hpp
