+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/include/nova_behavior_tree/condition/cube_detected_condition.hpp
