+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/plugins/archive/check_detected_condition.cpp
