+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/plugins/archive/compute_search_goals_condition.cpp
