+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/plugins/archive/first_goal_reached_condition.cpp
