+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/plugins/archive/get_first_goal_condition.cpp
