+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/plugins/archive/goals_empty_condition.cpp
