+++ src/ros/rover/nav2_autonomous/nova_behavior_tree/plugins/archive/query_position_condition.cpp
